Fix variable declaration and shadowing issues
