+++ src/extension/outputChannelLogger.ts
@@ -11,43 +11,6 @@ import { SettingsHelper } from "./settingsHelper";
 import { OutputChannel } from "vscode";
 import * as vscode from "vscode";
 
-export class DelayedOutputChannelLogger implements ILogger {
-    private outputChannelLogger: OutputChannelLogger;
-
-    constructor(private channelName: string) { }
-
-    public logInternalMessage(logLevel: LogLevel, message: string) {
-        this.logger.logInternalMessage(logLevel, message);
-    }
-
-    public logMessage(message: string, formatMessage: boolean = true) {
-        this.logger.logMessage(message, formatMessage);
-    }
-
-    public logError(errorMessage: string, error?: any, logStack: boolean = true) {
-        this.logger.logError(errorMessage, error, logStack);
-    }
-
-    public logStreamData(data: Buffer, stream: NodeJS.WritableStream) {
-        this.logger.logStreamData(data, stream);
-    }
-
-    public logString(data: string) {
-        this.logger.logString(data);
-    }
-
-    public setFocusOnLogChannel() {
-        this.logger.setFocusOnLogChannel();
-    }
-
-    private get logger(): OutputChannelLogger {
-        if (!this.outputChannelLogger) {
-            this.outputChannelLogger = new OutputChannelLogger(vscode.window.createOutputChannel(this.channelName));
-        }
-        return this.outputChannelLogger;
-    }
-}
-
 export class OutputChannelLogger implements ILogger {
     private outputChannel: OutputChannel;
 
@@ -96,5 +59,42 @@ export class OutputChannelLogger implements ILogger {
 
     private getFormattedInternalMessage(logLevel: LogLevel, message: string) {
         return (`${LogHelper.INTERNAL_TAG} [${LogLevel[logLevel]}] ${message}`);
+    }
+}
+
+export class DelayedOutputChannelLogger implements ILogger {
+    private outputChannelLogger: OutputChannelLogger;
+
+    constructor(private channelName: string) { }
+
+    public logInternalMessage(logLevel: LogLevel, message: string) {
+        this.logger.logInternalMessage(logLevel, message);
+    }
+
+    public logMessage(message: string, formatMessage: boolean = true) {
+        this.logger.logMessage(message, formatMessage);
+    }
+
+    public logError(errorMessage: string, error?: any, logStack: boolean = true) {
+        this.logger.logError(errorMessage, error, logStack);
+    }
+
+    public logStreamData(data: Buffer, stream: NodeJS.WritableStream) {
+        this.logger.logStreamData(data, stream);
+    }
+
+    public logString(data: string) {
+        this.logger.logString(data);
+    }
+
+    public setFocusOnLogChannel() {
+        this.logger.setFocusOnLogChannel();
+    }
+
+    private get logger(): OutputChannelLogger {
+        if (!this.outputChannelLogger) {
+            this.outputChannelLogger = new OutputChannelLogger(vscode.window.createOutputChannel(this.channelName));
+        }
+        return this.outputChannelLogger;
     }
 }
